No separate transport in a worker thread
