+++ inline-js-core/jsbits/eval.mjs
@@ -1,20 +1,13 @@
 import module from "module";
-import path from "path";
 import process from "process";
 import { StringDecoder } from "string_decoder";
-import url from "url";
 import vm from "vm";
-import { Worker } from "worker_threads";
 import JSVal from "./jsval.mjs";
+import { pipeRead, pipeWrite } from "./pipe.mjs";
 
 function errorStringify(err) {
   return err.stack ? err.stack : `${err}`;
 }
-
-process.on("uncaughtException", err => {
-  process.stderr.write(errorStringify(err));
-  throw err;
-});
 
 process.on("unhandledRejection", err => {
   process.stderr.write(errorStringify(err));
@@ -22,18 +15,8 @@ process.on("unhandledRejection", err => {
 });
 
 const decoder = new StringDecoder("utf8"),
-  ipc = new Worker(
-    path.join(
-      path.dirname(url.fileURLToPath(import.meta.url)),
-      "transport.mjs"
-    ),
-    {
-      workerData: [
-        Number.parseInt(process.argv[process.argv.length - 1]),
-        Number.parseInt(process.argv[process.argv.length - 2])
-      ]
-    }
-  );
+  node_read_fd = Number.parseInt(process.argv[process.argv.length - 1]),
+  node_write_fd = Number.parseInt(process.argv[process.argv.length - 2]);
 
 function bufferFromU32(x) {
   const buf = Buffer.allocUnsafe(4);
@@ -44,9 +27,21 @@ function bufferFromU32(x) {
 global.JSVal = JSVal;
 global.require = module.createRequire(import.meta.url);
 
-ipc.on("message", async ([msg_id, msg_tag, buf]) => {
+function postHostMessage(msg_id, is_err, result) {
+  const result_buf = Buffer.from(result),
+    msg_buf = Buffer.allocUnsafe(12 + result_buf.length);
+  msg_buf.writeUInt32LE(8 + result_buf.length, 0);
+  msg_buf.writeUInt32LE(msg_id, 4);
+  msg_buf.writeUInt32LE(Number(is_err), 8);
+  result_buf.copy(msg_buf, 12);
+  pipeWrite(node_write_fd, msg_buf);
+}
+
+async function handleHostMessage(msg_buf) {
+  const msg_id = msg_buf.readUInt32LE(0),
+    msg_tag = msg_buf.readUInt32LE(4),
+    buf = msg_buf.slice(8);
   try {
-    buf = Buffer.from(buf);
     switch (msg_tag) {
       case 0: {
         const ret_tag = buf.readUInt32LE(0),
@@ -60,7 +55,7 @@ ipc.on("message", async ([msg_id, msg_tag, buf]) => {
             eval_result === undefined ? null : eval_result
           ),
           promise_result = await promise;
-        ipc.postMessage([
+        postHostMessage(
           msg_id,
           false,
           ret_tag === 1
@@ -68,11 +63,11 @@ ipc.on("message", async ([msg_id, msg_tag, buf]) => {
             : ret_tag === 2
             ? Buffer.allocUnsafe(0)
             : promise_result
-        ]);
+        );
         break;
       }
       case 1: {
-        ipc.postMessage([msg_id, false, bufferFromU32(JSVal.newJSVal(buf))]);
+        postHostMessage(msg_id, false, bufferFromU32(JSVal.newJSVal(buf)));
         break;
       }
       default: {
@@ -80,6 +75,18 @@ ipc.on("message", async ([msg_id, msg_tag, buf]) => {
       }
     }
   } catch (err) {
-    ipc.postMessage([msg_id, true, errorStringify(err)]);
+    postHostMessage(msg_id, true, errorStringify(err));
   }
-});
+}
+
+async function onHostMessage() {
+  const msg_len_buf = Buffer.allocUnsafe(4);
+  await pipeRead(node_read_fd, msg_len_buf, 0, 4);
+  const msg_len = msg_len_buf.readUInt32LE(0),
+    msg_buf = Buffer.allocUnsafe(msg_len);
+  await pipeRead(node_read_fd, msg_buf, 0, msg_len);
+  setImmediate(handleHostMessage, msg_buf);
+  setImmediate(onHostMessage);
+}
+
+onHostMessage();
